+++ src/ailr/__init__.py
@@ -42,7 +42,7 @@ from ailr.tasks import (
     ScreeningTask,
 )
 
-__version__ = "0.30.0"
+__version__ = "0.31.0"
 
 __all__ = [
     "Project",